+++ src/Console/ConsoleEvent.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: marco.bunge
+ * Date: 05.02.2017
+ * Time: 11:47
+ */
+
+namespace Console;
+
+
+class ConsoleEvent
+{
+
+}